+++ CHANGELOG.md
@@ -147,3 +147,4 @@ can get data for both organization and organization brand pages.
 * Deprecated V2 connection classes
 
 ## 4.0.1 (Work in progress)
+* Add boolean support to JsonUtils