+++ CHANGELOG.md
@@ -147,4 +147,5 @@ can get data for both organization and organization brand pages.
 * Deprecated V2 connection classes
 
 ## 4.0.1 (Work in progress)
+* Better support for non-default versioned month
 * Add boolean support to JsonUtils